requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
neo4j==4.1.1
pylint==2.6.0
python-dotenv==0.14.0
pytz==2020.1
ratelimit==2.2.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.11
wrapt==1.12.1
